Move to akka actors Cleanup/seperate protocol Added basic slf4j/logback config
