+++ project/plugins.sbt
@@ -5,3 +5,5 @@ addSbtPlugin("com.timushev.sbt" % "sbt-updates" % "0.1.6")
 addSbtPlugin("net.virtual-void" % "sbt-dependency-graph" % "0.7.4")
 
 addSbtPlugin("com.typesafe.sbt" % "sbt-scalariform" % "1.3.0")
+
+addSbtPlugin("com.github.mpeltonen" % "sbt-idea" % "1.6.0")
