+++ src/main/resources/application.conf
@@ -1,0 +1,4 @@
+akka {
+  loggers = ["akka.event.slf4j.Slf4jLogger"]
+  loglevel = "INFO"
+}
