+++ src/main/scala/org/ensime/config/ProjectConfig.scala
@@ -733,7 +733,7 @@ object ProjectConfig {
   }
 }
 
-class ReplConfig(val classpath: String) {}
+class ReplConfig(val classpath: String)
 
 case class ProjectConfig(
     name: Option[String] = None,
