+++ src/main/scala/org/ensime/protocol/ConnectionInfo.scala
@@ -1,0 +1,7 @@
+package org.ensime.protocol
+
+class ConnectionInfo {
+  val pid = None
+  val serverName: String = "ENSIME-ReferenceServer"
+  val protocolVersion: String = "0.8.9"
+}
