+++ src/main/scala/org/ensime/protocol/SwankProtocol.scala
@@ -1,31 +1,26 @@
 package org.ensime.protocol
 
 import java.io._
-import org.ensime.config.{ ProjectConfig, ReplConfig }
-import org.ensime.indexer.MethodBytecode
+import akka.actor.ActorRef
+import org.ensime.config.ProjectConfig
 import org.ensime.model._
 import org.ensime.server._
 import org.ensime.util._
 import org.ensime.util.SExp._
-import scala.actors._
-import scala.tools.nsc.io.ZipArchive
-import scala.tools.nsc.util.{ Position, RangePosition }
 import scala.util.parsing.input
 
-object SwankProtocol extends SwankProtocol {}
-
-trait SwankProtocol extends Protocol {
-
-  class ConnectionInfo {
-    val pid = None
-    val serverName: String = "ENSIME-ReferenceServer"
-    val protocolVersion: String = "0.8.7"
-  }
+class SwankProtocol extends Protocol {
 
   /**
-   * Protocol Version: 0.8.7
+   * Protocol Version: 0.8.9
    *
    * Protocol Change Log:
+   *   0.8.9
+   *     Remove Incremental builder - removed
+   *       swank:builder-init
+   *       swank:builder-update-files
+   *       swank:builder-add-files
+   *       swank:builder-remove-files
    *   0.8.8
    *     Add optional :archive member to Position and RangePosition
    *   0.8.7
@@ -72,12 +67,14 @@ trait SwankProtocol extends Protocol {
 
   import ProtocolConst._
 
-  private var outPeer: Actor = null
+  override val conversions: ProtocolConversions = new SwankProtocolConversions
+
+  private var outPeer: ActorRef = null
   private var rpcTarget: RPCTarget = null
 
   def peer = outPeer
 
-  def setOutputActor(peer: Actor) { outPeer = peer }
+  def setOutputActor(peer: ActorRef) { outPeer = peer }
 
   def setRPCTarget(target: RPCTarget) { this.rpcTarget = target }
 
@@ -95,7 +92,7 @@ trait SwankProtocol extends Protocol {
 
   private def fillArray(in: java.io.Reader, a: Array[Char]) {
     var n = 0
-    var l = a.length
+    val l = a.length
     var charsRead = 0
     while (n < l) {
       charsRead = in.read(a, n, l - n)
@@ -518,8 +515,7 @@ trait SwankProtocol extends Protocol {
 
     println("\nHandling RPC: " + form.toReadableString(debug = true))
 
-    def oops = sendRPCError(ErrMalformedRPC,
-      Some("Malformed " + callType + " call: " + form), callId)
+    def oops() = sendRPCError(ErrMalformedRPC, Some("Malformed " + callType + " call: " + form), callId)
 
     callType match {
 
@@ -550,7 +546,7 @@ trait SwankProtocol extends Protocol {
        *   :version "0.7")) 42)
        */
       case "swank:connection-info" =>
-        sendRPCReturn(toWF(new ConnectionInfo()), callId)
+        sendRPCReturn(conversions.toWF(new ConnectionInfo()), callId)
 
       /**
        * Doc RPC:
@@ -582,11 +578,9 @@ trait SwankProtocol extends Protocol {
           case SExpList(head :: (conf: SExpList) :: body) =>
             ProjectConfig.fromSExp(conf) match {
               case Right(config) => rpcTarget.rpcInitProject(config, callId)
-              case Left(t) => sendRPCError(ErrExceptionInRPC,
-                Some(t.toString),
-                callId)
+              case Left(t) => sendRPCError(ErrExceptionInRPC, Some(t.toString), callId)
             }
-          case _ => oops
+          case _ => oops()
         }
 
       /**
@@ -636,7 +630,7 @@ trait SwankProtocol extends Protocol {
         form match {
           case SExpList(head :: (IntAtom(id)) :: body) =>
             rpcTarget.rpcExecUndo(id, callId)
-          case _ => oops
+          case _ => oops()
         }
 
       /**
@@ -660,95 +654,6 @@ trait SwankProtocol extends Protocol {
 
       /**
        * Doc RPC:
-       *   swank:builder-init
-       * Summary:
-       *   Initialize the incremental builder and kick off a full rebuild.
-       * Arguments:
-       *   None
-       * Return:
-       *   None
-       * Example call:
-       *   (:swank-rpc (swank:builder-init) 42)
-       * Example return:
-       *   (:return (:ok ()) 42)
-       */
-      case "swank:builder-init" =>
-        rpcTarget.rpcBuilderInit(callId)
-
-      /**
-       * Doc RPC:
-       *   swank:builder-update-files
-       * Summary:
-       *   Signal to the incremental builder that the given files
-       *   have changed and must be rebuilt. Triggers rebuild.
-       * Arguments:
-       *   List of Strings:Filenames, absolute or relative to the project root.
-       * Return:
-       *   None
-       * Example call:
-       *   (:swank-rpc (swank:builder-update-files
-       *   ("/ensime/src/main/scala/org/ensime/server/Analyzer.scala")) 42)
-       * Example return:
-       *   (:return (:ok ()) 42)
-       */
-      case "swank:builder-update-files" =>
-        form match {
-          case SExpList(head :: SExpList(filenames) :: body) =>
-            val files = filenames.map(_.toString)
-            rpcTarget.rpcBuilderUpdateFiles(files, callId)
-          case _ => oops
-        }
-
-      /**
-       * Doc RPC:
-       *   swank:builder-add-files
-       * Summary:
-       *   Signal to the incremental builder that the given files
-       *   should be added to the build. Triggers rebuild.
-       * Arguments:
-       *   List of Strings:Filenames, absolute or relative to the project root.
-       * Return:
-       *   None
-       * Example call:
-       *   (:swank-rpc (swank:builder-add-files
-       *   ("/ensime/src/main/scala/org/ensime/server/Analyzer.scala")) 42)
-       * Example return:
-       *   (:return (:ok ()) 42)
-       */
-      case "swank:builder-add-files" =>
-        form match {
-          case SExpList(head :: SExpList(filenames) :: body) =>
-            val files = filenames.map(_.toString)
-            rpcTarget.rpcBuilderAddFiles(files, callId)
-          case _ => oops
-        }
-
-      /**
-       * Doc RPC:
-       *   swank:builder-remove-files
-       * Summary:
-       *   Signal to the incremental builder that the given files
-       *   should be removed from the build. Triggers rebuild.
-       * Arguments:
-       *   List of Strings:Filenames, absolute or relative to the project root.
-       * Return:
-       *   None
-       * Example call:
-       *   (:swank-rpc (swank:builder-remove-files
-       *   ("/ensime/src/main/scala/org/ensime/server/Analyzer.scala")) 42)
-       * Example return:
-       *   (:return (:ok ()) 42)
-       */
-      case "swank:builder-remove-files" =>
-        form match {
-          case SExpList(head :: SExpList(filenames) :: body) =>
-            val files = filenames.map(_.toString)
-            rpcTarget.rpcBuilderRemoveFiles(files, callId)
-          case _ => oops
-        }
-
-      /**
-       * Doc RPC:
        *   swank:remove-file
        * Summary:
        *   Remove a file from consideration by the ENSIME analyzer.
@@ -765,7 +670,7 @@ trait SwankProtocol extends Protocol {
         form match {
           case SExpList(head :: StringAtom(file) :: body) =>
             rpcTarget.rpcRemoveFile(file, callId)
-          case _ => oops
+          case _ => oops()
         }
 
       /**
@@ -789,7 +694,7 @@ trait SwankProtocol extends Protocol {
             rpcTarget.rpcTypecheckFiles(List(SourceFileInfo(new File(file), Some(contents))), callId)
           case SExpList(head :: StringAtom(file) :: body) =>
             rpcTarget.rpcTypecheckFiles(List(SourceFileInfo(file)), callId)
-          case _ => oops
+          case _ => oops()
         }
 
       /**
@@ -811,7 +716,7 @@ trait SwankProtocol extends Protocol {
           case SExpList(head :: SExpList(strings) :: body) =>
             val filenames = strings.collect { case StringAtom(s) => SourceFileInfo(s) }.toList
             rpcTarget.rpcTypecheckFiles(filenames, callId)
-          case _ => oops
+          case _ => oops()
         }
 
       /**
@@ -844,7 +749,7 @@ trait SwankProtocol extends Protocol {
               case _ => PatchInsert(0, "")
             }.toList
             rpcTarget.rpcPatchSource(file, edits, callId)
-          case _ => oops
+          case _ => oops()
         }
 
       /**
@@ -886,7 +791,7 @@ trait SwankProtocol extends Protocol {
           case SExpList(head :: SExpList(filenames) :: body) =>
             val files = filenames.map(_.toString)
             rpcTarget.rpcFormatFiles(files, callId)
-          case _ => oops
+          case _ => oops()
         }
 
       /**
@@ -911,7 +816,7 @@ trait SwankProtocol extends Protocol {
           case SExpList(head :: SExpList(keywords) :: IntAtom(maxResults) :: body) =>
             rpcTarget.rpcPublicSymbolSearch(
               keywords.map(_.toString).toList, maxResults, callId)
-          case _ => oops
+          case _ => oops()
         }
 
       /**
@@ -934,7 +839,7 @@ trait SwankProtocol extends Protocol {
        *   "/ensime/src/main/scala/org/ensime/server/Analyzer.scala"
        *   2300 (Actor) 10) 42)
        * Example return:
-       *   (:return (:ok (((:name "scala.actors.Actor" :local-name "Actor"
+       *   (:return (:ok (((:name "akka.actor.Actor" :local-name "Actor"
        *   :decl-as trait :pos (:file "/lib/scala-library.jar" :offset -1)))))
        *   42)
        */
@@ -944,7 +849,7 @@ trait SwankProtocol extends Protocol {
             SExpList(names) :: IntAtom(maxResults) :: body) =>
             rpcTarget.rpcImportSuggestions(file, point,
               names.map(_.toString).toList, maxResults, callId)
-          case _ => oops
+          case _ => oops()
         }
 
       /**
@@ -979,7 +884,7 @@ trait SwankProtocol extends Protocol {
             BooleanAtom(reload) :: body) =>
             rpcTarget.rpcCompletionsAtPoint(
               file, point, maxResults, caseSens, reload, callId)
-          case _ => oops
+          case _ => oops()
         }
 
       /**
@@ -1002,7 +907,7 @@ trait SwankProtocol extends Protocol {
         form match {
           case SExpList(head :: StringAtom(path) :: StringAtom(prefix) :: body) =>
             rpcTarget.rpcPackageMemberCompletion(path, prefix, callId)
-          case _ => oops
+          case _ => oops()
         }
 
       /**
@@ -1031,7 +936,7 @@ trait SwankProtocol extends Protocol {
         form match {
           case SExpList(head :: IntAtom(id) :: body) =>
             rpcTarget.rpcCallCompletion(id, callId)
-          case _ => oops
+          case _ => oops()
         }
 
       /**
@@ -1056,7 +961,7 @@ trait SwankProtocol extends Protocol {
         form match {
           case SExpList(head :: StringAtom(file) :: IntAtom(point) :: body) =>
             rpcTarget.rpcUsesOfSymAtPoint(file, point, callId)
-          case _ => oops
+          case _ => oops()
         }
 
       /**
@@ -1079,7 +984,7 @@ trait SwankProtocol extends Protocol {
         form match {
           case SExpList(head :: IntAtom(id) :: body) =>
             rpcTarget.rpcTypeById(id, callId)
-          case _ => oops
+          case _ => oops()
         }
 
       /**
@@ -1101,7 +1006,7 @@ trait SwankProtocol extends Protocol {
         form match {
           case SExpList(head :: StringAtom(name) :: body) =>
             rpcTarget.rpcTypeByName(name, callId)
-          case _ => oops
+          case _ => oops()
         }
 
       /**
@@ -1127,7 +1032,7 @@ trait SwankProtocol extends Protocol {
           case SExpList(head :: StringAtom(name) :: StringAtom(file) ::
             OffsetRangeExtractor(range) :: body) =>
             rpcTarget.rpcTypeByNameAtPoint(name, file, range, callId)
-          case _ => oops
+          case _ => oops()
         }
 
       /**
@@ -1152,7 +1057,7 @@ trait SwankProtocol extends Protocol {
           case SExpList(head :: StringAtom(file) ::
             OffsetRangeExtractor(range) :: body) =>
             rpcTarget.rpcTypeAtPoint(file, range, callId)
-          case _ => oops
+          case _ => oops()
         }
 
       /**
@@ -1178,7 +1083,7 @@ trait SwankProtocol extends Protocol {
           case SExpList(head :: StringAtom(file) ::
             OffsetRangeExtractor(range) :: body) =>
             rpcTarget.rpcInspectTypeAtPoint(file, range, callId)
-          case _ => oops
+          case _ => oops()
         }
 
       /**
@@ -1201,7 +1106,7 @@ trait SwankProtocol extends Protocol {
         form match {
           case SExpList(head :: IntAtom(id) :: body) =>
             rpcTarget.rpcInspectTypeById(id, callId)
-          case _ => oops
+          case _ => oops()
         }
 
       /**
@@ -1225,7 +1130,7 @@ trait SwankProtocol extends Protocol {
         form match {
           case SExpList(head :: StringAtom(file) :: IntAtom(point) :: body) =>
             rpcTarget.rpcSymbolAtPoint(file, point, callId)
-          case _ => oops
+          case _ => oops()
         }
 
       /**
@@ -1249,7 +1154,7 @@ trait SwankProtocol extends Protocol {
         form match {
           case SExpList(head :: StringAtom(path) :: body) =>
             rpcTarget.rpcInspectPackageByPath(path, callId)
-          case _ => oops
+          case _ => oops()
         }
 
       /**
@@ -1289,7 +1194,7 @@ trait SwankProtocol extends Protocol {
             (params: SExp) :: BooleanAtom(interactive) :: body) =>
             rpcTarget.rpcPrepareRefactor(Symbol(tpe), procId,
               listOrEmpty(params).toSymbolMap, interactive, callId)
-          case _ => oops
+          case _ => oops()
         }
 
       /**
@@ -1313,7 +1218,7 @@ trait SwankProtocol extends Protocol {
         form match {
           case SExpList(dude :: IntAtom(procId) :: SymbolAtom(tpe) :: body) =>
             rpcTarget.rpcExecRefactor(Symbol(tpe), procId, callId)
-          case _ => oops
+          case _ => oops()
         }
 
       /**
@@ -1335,7 +1240,7 @@ trait SwankProtocol extends Protocol {
         form match {
           case SExpList(head :: IntAtom(procId) :: body) =>
             rpcTarget.rpcCancelRefactor(procId, callId)
-          case _ => oops
+          case _ => oops()
         }
 
       /**
@@ -1370,7 +1275,7 @@ trait SwankProtocol extends Protocol {
               tpe => Symbol(tpe.toString)).toList
             rpcTarget.rpcSymbolDesignations(filename, start,
               end, requestedTypes, callId)
-          case _ => oops
+          case _ => oops()
         }
 
       /**
@@ -1396,7 +1301,7 @@ trait SwankProtocol extends Protocol {
           case SExpList(head :: StringAtom(filename) :: IntAtom(start) ::
             IntAtom(end) :: body) =>
             rpcTarget.rpcExpandSelection(filename, start, end, callId)
-          case _ => oops
+          case _ => oops()
         }
 
       /**
@@ -1426,7 +1331,7 @@ trait SwankProtocol extends Protocol {
           case SExpList(head :: StringAtom(filename) ::
             IntAtom(line) :: body) =>
             rpcTarget.rpcMethodBytecode(filename, line, callId)
-          case _ => oops
+          case _ => oops()
         }
 
       /**
@@ -1447,7 +1352,7 @@ trait SwankProtocol extends Protocol {
         form match {
           case SExpList(head :: body) =>
             rpcTarget.rpcDebugActiveVM(callId)
-          case _ => oops
+          case _ => oops()
         }
 
       /**
@@ -1469,7 +1374,7 @@ trait SwankProtocol extends Protocol {
         form match {
           case SExpList(head :: StringAtom(commandLine) :: body) =>
             rpcTarget.rpcDebugStartVM(commandLine, callId)
-          case _ => oops
+          case _ => oops()
         }
 
       /**
@@ -1491,7 +1396,7 @@ trait SwankProtocol extends Protocol {
         form match {
           case SExpList(head :: StringAtom(hostname) :: StringAtom(port) :: body) =>
             rpcTarget.rpcDebugAttachVM(hostname, port, callId)
-          case _ => oops
+          case _ => oops()
         }
 
       /**
@@ -1512,7 +1417,7 @@ trait SwankProtocol extends Protocol {
         form match {
           case SExpList(head :: body) =>
             rpcTarget.rpcDebugStopVM(callId)
-          case _ => oops
+          case _ => oops()
         }
 
       /**
@@ -1535,7 +1440,7 @@ trait SwankProtocol extends Protocol {
           case SExpList(head :: StringAtom(filename) ::
             IntAtom(line) :: body) =>
             rpcTarget.rpcDebugBreak(filename, line, callId)
-          case _ => oops
+          case _ => oops()
         }
 
       /**
@@ -1558,7 +1463,7 @@ trait SwankProtocol extends Protocol {
           case SExpList(head :: StringAtom(filename) ::
             IntAtom(line) :: body) =>
             rpcTarget.rpcDebugClearBreak(filename, line, callId)
-          case _ => oops
+          case _ => oops()
         }
 
       /**
@@ -1579,7 +1484,7 @@ trait SwankProtocol extends Protocol {
         form match {
           case SExpList(head :: body) =>
             rpcTarget.rpcDebugClearAllBreaks(callId)
-          case _ => oops
+          case _ => oops()
         }
 
       /**
@@ -1601,7 +1506,7 @@ trait SwankProtocol extends Protocol {
         form match {
           case SExpList(head :: body) =>
             rpcTarget.rpcDebugListBreaks(callId)
-          case _ => oops
+          case _ => oops()
         }
 
       /**
@@ -1622,7 +1527,7 @@ trait SwankProtocol extends Protocol {
         form match {
           case SExpList(head :: body) =>
             rpcTarget.rpcDebugRun(callId)
-          case _ => oops
+          case _ => oops()
         }
 
       /**
@@ -1643,7 +1548,7 @@ trait SwankProtocol extends Protocol {
         form match {
           case SExpList(head :: StringAtom(threadId) :: body) =>
             rpcTarget.rpcDebugContinue(threadId.toLong, callId)
-          case _ => oops
+          case _ => oops()
         }
 
       /**
@@ -1665,7 +1570,7 @@ trait SwankProtocol extends Protocol {
         form match {
           case SExpList(head :: StringAtom(threadId) :: body) =>
             rpcTarget.rpcDebugStep(threadId.toLong, callId)
-          case _ => oops
+          case _ => oops()
         }
 
       /**
@@ -1687,7 +1592,7 @@ trait SwankProtocol extends Protocol {
         form match {
           case SExpList(head :: StringAtom(threadId) :: body) =>
             rpcTarget.rpcDebugNext(threadId.toLong, callId)
-          case _ => oops
+          case _ => oops()
         }
 
       /**
@@ -1709,7 +1614,7 @@ trait SwankProtocol extends Protocol {
         form match {
           case SExpList(head :: StringAtom(threadId) :: body) =>
             rpcTarget.rpcDebugStepOut(threadId.toLong, callId)
-          case _ => oops
+          case _ => oops()
         }
 
       /**
@@ -1732,7 +1637,7 @@ trait SwankProtocol extends Protocol {
         form match {
           case SExpList(head :: StringAtom(threadId) :: StringAtom(name) :: body) =>
             rpcTarget.rpcDebugLocateName(threadId.toLong, name, callId)
-          case _ => oops
+          case _ => oops()
         }
 
       /**
@@ -1755,7 +1660,7 @@ trait SwankProtocol extends Protocol {
         form match {
           case SExpList(head :: DebugLocationExtractor(loc) :: body) =>
             rpcTarget.rpcDebugValue(loc, callId)
-          case _ => oops
+          case _ => oops()
         }
 
       /**
@@ -1779,7 +1684,7 @@ trait SwankProtocol extends Protocol {
         form match {
           case SExpList(head :: StringAtom(threadId) :: DebugLocationExtractor(loc) :: body) =>
             rpcTarget.rpcDebugToString(threadId.toLong, loc, callId)
-          case _ => oops
+          case _ => oops()
         }
 
       /**
@@ -1803,7 +1708,7 @@ trait SwankProtocol extends Protocol {
           case SExpList(head :: DebugLocationExtractor(loc) ::
             StringAtom(newValue) :: body) =>
             rpcTarget.rpcDebugSetValue(loc, newValue, callId)
-          case _ => oops
+          case _ => oops()
         }
 
       /**
@@ -1828,7 +1733,7 @@ trait SwankProtocol extends Protocol {
           case SExpList(head :: StringAtom(threadId) ::
             IntAtom(index) :: IntAtom(count) :: body) =>
             rpcTarget.rpcDebugBacktrace(threadId.toLong, index, count, callId)
-          case _ => oops
+          case _ => oops()
         }
 
       /**
@@ -1902,46 +1807,6 @@ trait SwankProtocol extends Protocol {
         detail.map(strToSExp).getOrElse(NilAtom())))
   }
 
-  object SExpConversion {
-
-    implicit def posToSExp(pos: Position): SExp = {
-      if (pos.isDefined) {
-        val underlying = pos.source.file.underlyingSource.orNull
-        val archive: SExp = underlying match {
-          case a: ZipArchive if a != pos.source.file => a.path
-          case _ => 'nil
-        }
-        SExp.propList(
-          (":file", pos.source.path),
-          (":archive", archive),
-          (":offset", pos.point))
-      } else {
-        'nil
-      }
-    }
-
-    implicit def posToSExp(pos: RangePosition): SExp = {
-      if (pos.isDefined) {
-        val underlying = pos.source.file.underlyingSource.orNull
-        val archive: SExp = underlying match {
-          case a: ZipArchive if a != pos.source.file => a.path
-          case _ => 'nil
-        }
-        SExp.propList(
-          (":file", pos.source.path),
-          (":archive", archive),
-          (":offset", pos.point),
-          (":start", pos.start),
-          (":end", pos.end))
-      } else {
-        'nil
-      }
-    }
-
-  }
-
-  import SExpConversion._
-
   object OffsetRangeExtractor {
     def unapply(sexp: SExp): Option[OffsetRange] = sexp match {
       case IntAtom(a) => Some(OffsetRange(a, a))
@@ -1983,709 +1848,6 @@ trait SwankProtocol extends Protocol {
         case _ => None
       }
     }
-  }
-
-  override def toWF(obj: DebugLocation): SExp = {
-    obj match {
-      case obj: DebugObjectReference => toWF(obj)
-      case obj: DebugArrayElement => toWF(obj)
-      case obj: DebugObjectField => toWF(obj)
-      case obj: DebugStackSlot => toWF(obj)
-    }
-  }
-  def toWF(obj: DebugObjectReference): SExp = {
-    SExp(
-      key(":type"), 'reference,
-      key(":object-id"), obj.objectId.toString)
-  }
-  def toWF(obj: DebugArrayElement): SExp = {
-    SExp(
-      key(":type"), 'element,
-      key(":object-id"), obj.objectId.toString,
-      key(":index"), obj.index)
-  }
-  def toWF(obj: DebugObjectField): SExp = {
-    SExp(
-      key(":type"), 'field,
-      key(":object-id"), obj.objectId.toString,
-      key(":field"), obj.name)
-  }
-  def toWF(obj: DebugStackSlot): SExp = {
-    SExp(
-      key(":type"), 'slot,
-      key(":thread-id"), obj.threadId.toString,
-      key(":frame"), obj.frame,
-      key(":offset"), obj.offset)
-  }
-
-  override def toWF(obj: DebugValue): SExp = {
-    obj match {
-      case obj: DebugPrimitiveValue => toWF(obj)
-      case obj: DebugObjectInstance => toWF(obj)
-      case obj: DebugArrayInstance => toWF(obj)
-      case obj: DebugStringInstance => toWF(obj)
-      case obj: DebugNullValue => toWF(obj)
-    }
-  }
-
-  override def toWF(obj: DebugNullValue): SExp = {
-    SExp(
-      key(":val-type"), 'null,
-      key(":type-name"), obj.typeName)
-  }
-
-  override def toWF(obj: DebugPrimitiveValue): SExp = {
-    SExp(
-      key(":val-type"), 'prim,
-      key(":summary"), obj.summary,
-      key(":type-name"), obj.typeName)
-  }
-
-  override def toWF(obj: DebugClassField): SExp = {
-    SExp(
-      key(":index"), obj.index,
-      key(":name"), obj.name,
-      key(":summary"), obj.summary,
-      key(":type-name"), obj.typeName)
-  }
-
-  override def toWF(obj: DebugObjectInstance): SExp = {
-    SExp(
-      key(":val-type"), 'obj,
-      key(":fields"), SExpList(obj.fields.map(toWF)),
-      key(":type-name"), obj.typeName,
-      key(":object-id"), obj.objectId.toString)
-  }
-
-  override def toWF(obj: DebugStringInstance): SExp = {
-    SExp(
-      key(":val-type"), 'str,
-      key(":summary"), obj.summary,
-      key(":fields"), SExpList(obj.fields.map(toWF)),
-      key(":type-name"), obj.typeName,
-      key(":object-id"), obj.objectId.toString)
-  }
-
-  override def toWF(obj: DebugArrayInstance): SExp = {
-    SExp(
-      key(":val-type"), 'arr,
-      key(":length"), obj.length,
-      key(":type-name"), obj.typeName,
-      key(":element-type-name"), obj.elementTypeName,
-      key(":object-id"), obj.objectId.toString)
-  }
-
-  override def toWF(obj: DebugStackLocal): SExp = {
-    SExp(
-      key(":index"), obj.index,
-      key(":name"), obj.name,
-      key(":summary"), obj.summary,
-      key(":type-name"), obj.typeName)
-  }
-
-  override def toWF(obj: DebugStackFrame): SExp = {
-    SExp(
-      key(":index"), obj.index,
-      key(":locals"), SExpList(obj.locals.map(toWF)),
-      key(":num-args"), obj.numArguments,
-      key(":class-name"), obj.className,
-      key(":method-name"), obj.methodName,
-      key(":pc-location"), toWF(obj.pcLocation),
-      key(":this-object-id"), obj.thisObjectId.toString)
-  }
-
-  override def toWF(obj: DebugBacktrace): SExp = {
-    SExp(
-      key(":frames"), SExpList(obj.frames.map(toWF)),
-      key(":thread-id"), obj.threadId.toString,
-      key(":thread-name"), obj.threadName)
-  }
-
-  override def toWF(pos: SourcePosition): SExp = {
-    SExp(
-      key(":file"), pos.file.getAbsolutePath,
-      key(":line"), pos.line)
-  }
-
-  def toWF(info: ConnectionInfo): SExp = {
-    SExp(
-      key(":pid"), 'nil,
-      key(":implementation"),
-      SExp(key(":name"), info.serverName),
-      key(":version"), info.protocolVersion)
-  }
-
-  override def toWF(evt: SendBackgroundMessageEvent): SExp = {
-    SExp(key(":background-message"), evt.code,
-      evt.detail.map(strToSExp).getOrElse(NilAtom()))
-  }
-
-  /**
-   * Doc Event:
-   *   :compiler-ready
-   * Summary:
-   *   Signal that the compiler has finished its initial compilation and the server
-   *   is ready to accept RPC calls.
-   * Structure:
-   *   (:compiler-ready)
-   */
-  def toWF(evt: AnalyzerReadyEvent): SExp = {
-    SExp(key(":compiler-ready"))
-  }
-
-  /**
-   * Doc Event:
-   *   :full-typecheck-finished
-   * Summary:
-   *   Signal that the compiler has finished compilation of the entire project.
-   * Structure:
-   *   (:full-typecheck-finished)
-   */
-  def toWF(evt: FullTypeCheckCompleteEvent): SExp = {
-    SExp(key(":full-typecheck-finished"))
-  }
-
-  /**
-   * Doc Event:
-   *   :indexer-ready
-   * Summary:
-   *   Signal that the indexer has finished indexing the classpath.
-   * Structure:
-   *   (:indexer-ready)
-   */
-  def toWF(evt: IndexerReadyEvent): SExp = {
-    SExp(key(":indexer-ready"))
-  }
-
-  /**
-   * Doc Event:
-   *   :scala-notes
-   * Summary:
-   *   Notify client when Scala compiler generates errors,warnings or other notes.
-   * Structure:
-   *   (:scala-notes
-   *   notes //List of Note
-   *   )
-   */
-  //-----------------------
-  /**
-   * Doc Event:
-   *   :java-notes
-   * Summary:
-   *   Notify client when Java compiler generates errors,warnings or other notes.
-   * Structure:
-   *   (:java-notes
-   *   notes //List of Note
-   *   )
-   */
-  def toWF(evt: NewNotesEvent): SExp = {
-    if (evt.lang == 'scala) SExp(key(":scala-notes"), toWF(evt.notelist))
-    else SExp(key(":java-notes"), toWF(evt.notelist))
-  }
-
-  /**
-   * Doc Event:
-   *   :clear-all-scala-notes
-   * Summary:
-   *   Notify client when Scala notes have become invalidated. Editor should consider
-   *   all Scala related notes to be stale at this point.
-   * Structure:
-   *   (:clear-all-scala-notes)
-   */
-  //-----------------------
-  /**
-   * Doc Event:
-   *   :clear-all-java-notes
-   * Summary:
-   *   Notify client when Java notes have become invalidated. Editor should consider
-   *   all Java related notes to be stale at this point.
-   * Structure:
-   *   (:clear-all-java-notes)
-   */
-  def toWF(evt: ClearAllNotesEvent): SExp = {
-    if (evt.lang == 'scala) SExp(key(":clear-all-scala-notes"))
-    else SExp(key(":clear-all-java-notes"))
-  }
-
-  def toWF(evt: DebugEvent): SExp = {
-    evt match {
-      /**
-       * Doc Event:
-       *   :debug-event (:type output)
-       * Summary:
-       *   Communicates stdout/stderr of debugged VM to client.
-       * Structure:
-       *   (:debug-event
-       *     (:type //Symbol: output
-       *      :body //String: A chunk of output text
-       *   ))
-       */
-      case DebugOutputEvent(out: String) =>
-        SExp(key(":debug-event"),
-          SExp(key(":type"), 'output,
-            key(":body"), out))
-
-      /**
-       * Doc Event:
-       *   :debug-event (:type step)
-       * Summary:
-       *   Signals that the debugged VM has stepped to a new location and is now
-       *     paused awaiting control.
-       * Structure:
-       *   (:debug-event
-       *     (:type //Symbol: step
-       *      :thread-id //String: The unique thread id of the paused thread.
-       *      :thread-name //String: The informal name of the paused thread.
-       *      :file //String: The source file the VM stepped into.
-       *      :line //Int: The source line the VM stepped to.
-       *   ))
-       */
-      case DebugStepEvent(threadId, threadName, pos) =>
-        SExp(key(":debug-event"),
-          SExp(key(":type"), 'step,
-            key(":thread-id"), threadId.toString,
-            key(":thread-name"), threadName,
-            key(":file"), pos.file.getAbsolutePath,
-            key(":line"), pos.line))
-
-      /**
-       * Doc Event:
-       *   :debug-event (:type breakpoint)
-       * Summary:
-       *   Signals that the debugged VM has stopped at a breakpoint.
-       * Structure:
-       *   (:debug-event
-       *     (:type //Symbol: breakpoint
-       *      :thread-id //String: The unique thread id of the paused thread.
-       *      :thread-name //String: The informal name of the paused thread.
-       *      :file //String: The source file the VM stepped into.
-       *      :line //Int: The source line the VM stepped to.
-       *   ))
-       */
-      case DebugBreakEvent(threadId, threadName, pos) =>
-        SExp(key(":debug-event"),
-          SExp(key(":type"), 'breakpoint,
-            key(":thread-id"), threadId.toString,
-            key(":thread-name"), threadName,
-            key(":file"), pos.file.getAbsolutePath,
-            key(":line"), pos.line))
-
-      /**
-       * Doc Event:
-       *   :debug-event (:type death)
-       * Summary:
-       *   Signals that the debugged VM has exited.
-       * Structure:
-       *   (:debug-event
-       *     (:type //Symbol: death
-       *   ))
-       */
-      case DebugVMDeathEvent() =>
-        SExp(key(":debug-event"),
-          SExp(key(":type"), 'death))
-
-      /**
-       * Doc Event:
-       *   :debug-event (:type start)
-       * Summary:
-       *   Signals that the debugged VM has started.
-       * Structure:
-       *   (:debug-event
-       *     (:type //Symbol: start
-       *   ))
-       */
-      case DebugVMStartEvent() =>
-        SExp(key(":debug-event"),
-          SExp(key(":type"), 'start))
-
-      /**
-       * Doc Event:
-       *   :debug-event (:type disconnect)
-       * Summary:
-       *   Signals that the debugger has disconnected form the debugged VM.
-       * Structure:
-       *   (:debug-event
-       *     (:type //Symbol: disconnect
-       *   ))
-       */
-      case DebugVMDisconnectEvent() =>
-        SExp(key(":debug-event"),
-          SExp(key(":type"), 'disconnect))
-
-      /**
-       * Doc Event:
-       *   :debug-event (:type exception)
-       * Summary:
-       *   Signals that the debugged VM has thrown an exception and is now paused
-       *     waiting for control.
-       * Structure:
-       *   (:debug-event
-       *     (:type //Symbol: exception
-       *      :exception //String: The unique object id of the exception.
-       *      :thread-id //String: The unique thread id of the paused thread.
-       *      :thread-name //String: The informal name of the paused thread.
-       *      :file //String: The source file where the exception was caught,
-       *         or nil if no location is known.
-       *      :line //Int: The source line where the exception was thrown,
-       *         or nil if no location is known.
-       *   ))
-       */
-      case DebugExceptionEvent(excId, threadId, threadName, maybePos) =>
-        SExp(key(":debug-event"),
-          SExp(key(":type"), 'exception,
-            key(":exception"), excId.toString,
-            key(":thread-id"), threadId.toString,
-            key(":thread-name"), threadName,
-            key(":file"), maybePos.map { p =>
-              StringAtom(p.file.getAbsolutePath)
-            }.getOrElse('nil),
-            key(":line"), maybePos.map { p =>
-              IntAtom(p.line)
-            }.getOrElse('nil)))
-
-      /**
-       * Doc Event:
-       *   :debug-event (:type threadStart)
-       * Summary:
-       *   Signals that a new thread has started.
-       * Structure:
-       *   (:debug-event
-       *     (:type //Symbol: threadStart
-       *      :thread-id //String: The unique thread id of the new thread.
-       *   ))
-       */
-      case DebugThreadStartEvent(threadId: Long) =>
-        SExp(key(":debug-event"),
-          SExp(key(":type"), 'threadStart,
-            key(":thread-id"), threadId.toString))
-
-      /**
-       * Doc Event:
-       *   :debug-event (:type threadDeath)
-       * Summary:
-       *   Signals that a new thread has died.
-       * Structure:
-       *   (:debug-event
-       *     (:type //Symbol: threadDeath
-       *      :thread-id //String: The unique thread id of the new thread.
-       *   ))
-       */
-      case DebugThreadDeathEvent(threadId: Long) =>
-        SExp(key(":debug-event"),
-          SExp(key(":type"), 'threadDeath,
-            key(":thread-id"), threadId.toString))
-      case _ => SExp(key(":debug-event"))
-    }
-
-  }
-
-  def toWF(bp: Breakpoint): SExp = {
-    SExp(
-      key(":file"), bp.pos.file.getAbsolutePath,
-      key(":line"), bp.pos.line)
-  }
-
-  def toWF(bps: BreakpointList): SExp = {
-    SExp(
-      key(":active"), SExpList(bps.active.map { toWF }),
-      key(":pending"), SExpList(bps.pending.map { toWF }))
-  }
-
-  def toWF(config: ProjectConfig): SExp = {
-    SExp(
-      key(":project-name"), config.name.map(StringAtom).getOrElse('nil),
-      key(":source-roots"), SExp(
-        (config.sourceRoots ++ config.referenceSourceRoots).map {
-          f => StringAtom(f.getPath)
-        }))
-  }
-
-  def toWF(config: ReplConfig): SExp = {
-    SExp.propList((":classpath", strToSExp(config.classpath)))
-  }
-
-  def toWF(value: Boolean): SExp = {
-    if (value) TruthAtom()
-    else NilAtom()
-  }
-
-  def toWF(value: Null): SExp = {
-    NilAtom()
-  }
-
-  def toWF(value: String): SExp = {
-    StringAtom(value)
-  }
-
-  def toWF(note: Note): SExp = {
-    SExp(
-      key(":severity"), note.friendlySeverity,
-      key(":msg"), note.msg,
-      key(":beg"), note.beg,
-      key(":end"), note.end,
-      key(":line"), note.line,
-      key(":col"), note.col,
-      key(":file"), note.file)
-  }
-
-  def toWF(notelist: NoteList): SExp = {
-    val NoteList(isFull, notes) = notelist
-    SExp(
-      key(":is-full"),
-      toWF(isFull),
-      key(":notes"),
-      SExpList(notes.map(toWF)))
-  }
-
-  def toWF(values: Iterable[WireFormat]): SExp = {
-    SExpList(values.map(ea => ea.asInstanceOf[SExp]))
-  }
-
-  def toWF(value: CompletionSignature): SExp = {
-    SExp(
-      SExp(value.sections.map { section =>
-        SExpList(section.map { param =>
-          SExp(param._1, param._2)
-        })
-      }),
-      value.result)
-  }
-
-  def toWF(value: CompletionInfo): SExp = {
-    SExp.propList(
-      (":name", value.name),
-      (":type-sig", toWF(value.tpeSig)),
-      (":type-id", value.tpeId),
-      (":is-callable", value.isCallable),
-      (":to-insert", value.toInsert.map(strToSExp).getOrElse('nil)))
-  }
-
-  def toWF(value: CompletionInfoList): SExp = {
-    SExp.propList(
-      (":prefix", value.prefix),
-      (":completions", SExpList(value.completions.map(toWF))))
-  }
-
-  def toWF(value: PackageMemberInfoLight): SExp = {
-    SExp(key(":name"), value.name)
-  }
-
-  def toWF(value: SymbolInfo): SExp = {
-    SExp.propList(
-      (":name", value.name),
-      (":local-name", value.localName),
-      (":type", toWF(value.tpe)),
-      (":decl-pos", value.declPos),
-      (":is-callable", value.isCallable),
-      (":owner-type-id", value.ownerTypeId.map(intToSExp).getOrElse('nil)))
-  }
-
-  def toWF(value: FileRange): SExp = {
-    SExp.propList(
-      (":file", value.file),
-      (":start", value.start),
-      (":end", value.end))
-  }
-
-  def toWF(value: Position): SExp = {
-    posToSExp(value)
-  }
-
-  def toWF(value: RangePosition): SExp = {
-    posToSExp(value)
-  }
-
-  def toWF(value: NamedTypeMemberInfoLight): SExp = {
-    SExp.propList(
-      (":name", value.name),
-      (":type-sig", value.tpeSig),
-      (":type-id", value.tpeId),
-      (":is-callable", value.isCallable))
-  }
-
-  override def toWF(value: NamedTypeMemberInfo): SExp = {
-    SExp.propList(
-      (":name", value.name),
-      (":type", toWF(value.tpe)),
-      (":pos", value.pos),
-      (":decl-as", value.declaredAs))
-  }
-
-  override def toWF(value: EntityInfo): SExp = {
-    value match {
-      case value: PackageInfo => toWF(value)
-      case value: TypeInfo => toWF(value)
-      case value: NamedTypeMemberInfo => toWF(value)
-      case value: NamedTypeMemberInfoLight => toWF(value)
-      case unknownValue => throw new IllegalStateException("Unknown EntityInfo: " + unknownValue)
-    }
-  }
-
-  def toWF(value: TypeInfo): SExp = {
-    value match {
-      case value: ArrowTypeInfo =>
-        SExp.propList(
-          (":name", value.name),
-          (":type-id", value.id),
-          (":arrow-type", true),
-          (":result-type", toWF(value.resultType)),
-          (":param-sections", SExp(value.paramSections.map(toWF))))
-      case value: TypeInfo =>
-        SExp.propList((":name", value.name),
-          (":type-id", value.id),
-          (":full-name", value.fullName),
-          (":decl-as", value.declaredAs),
-          (":type-args", SExp(value.args.map(toWF))),
-          (":members", SExp(value.members.map(toWF))),
-          (":pos", value.pos),
-          (":outer-type-id", value.outerTypeId.map(intToSExp).getOrElse('nil)))
-      case unknownValue => throw new IllegalStateException("Unknown TypeInfo: " + unknownValue)
-    }
-  }
-
-  def toWF(value: PackageInfo): SExp = {
-    SExp.propList((":name", value.name),
-      (":info-type", 'package),
-      (":full-name", value.fullname),
-      (":members", SExpList(value.members.map(toWF))))
-  }
-
-  def toWF(value: CallCompletionInfo): SExp = {
-    SExp.propList(
-      (":result-type", toWF(value.resultType)),
-      (":param-sections", SExp(value.paramSections.map(toWF))))
-  }
-
-  def toWF(value: ParamSectionInfo): SExp = {
-    SExp.propList(
-      (":params", SExp(value.params.map {
-        case (nm, tp) => SExp(nm, toWF(tp))
-      })),
-      (":is-implicit", value.isImplicit))
-
-  }
-
-  def toWF(value: InterfaceInfo): SExp = {
-    SExp.propList(
-      (":type", toWF(value.tpe)),
-      (":via-view", value.viaView.map(strToSExp).getOrElse('nil)))
-  }
-
-  def toWF(value: TypeInspectInfo): SExp = {
-    SExp.propList(
-      (":type", toWF(value.tpe)),
-      (":info-type", 'typeInspect),
-      (":companion-id", value.companionId match {
-        case Some(id) => id
-        case None => 'nil
-      }), (":interfaces", SExp(value.supers.map(toWF))))
-  }
-
-  def toWF(value: RefactorFailure): SExp = {
-    SExp.propList(
-      (":procedure-id", value.procedureId),
-      (":status", 'failure),
-      (":reason", value.message))
-  }
-
-  def toWF(value: RefactorEffect): SExp = {
-    SExp.propList(
-      (":procedure-id", value.procedureId),
-      (":refactor-type", value.refactorType),
-      (":status", 'success),
-      (":changes", SExpList(value.changes.map(changeToWF))))
-  }
-
-  def toWF(value: RefactorResult): SExp = {
-    SExp.propList(
-      (":procedure-id", value.procedureId),
-      (":refactor-type", value.refactorType),
-      (":status", 'success),
-      (":touched-files", SExpList(value.touched.map(f => strToSExp(f.getAbsolutePath)))))
-  }
-
-  def toWF(value: SymbolSearchResults): SExp = {
-    SExpList(value.syms.map(toWF))
-  }
-
-  def toWF(value: ImportSuggestions): SExp = {
-    SExpList(value.symLists.map { l => SExpList(l.map(toWF)) })
-  }
-
-  private def toWF(pos: Option[(String, Int)]): SExp = {
-    pos match {
-      case Some((f, o)) => SExp.propList((":file", f), (":offset", o))
-      case _ => 'nil
-    }
-  }
-
-  def toWF(value: SymbolSearchResult): SExp = {
-    value match {
-      case value: TypeSearchResult =>
-        SExp.propList(
-          (":name", value.name),
-          (":local-name", value.localName),
-          (":decl-as", value.declaredAs),
-          (":pos", toWF(value.pos)))
-      case value: MethodSearchResult =>
-        SExp.propList(
-          (":name", value.name),
-          (":local-name", value.localName),
-          (":decl-as", value.declaredAs),
-          (":pos", toWF(value.pos)),
-          (":owner-name", value.owner))
-    }
-  }
-
-  def toWF(value: Undo): SExp = {
-    SExp.propList(
-      (":id", value.id),
-      (":changes", SExpList(value.changes.map(changeToWF))),
-      (":summary", value.summary))
-  }
-
-  def toWF(value: UndoResult): SExp = {
-    SExp.propList(
-      (":id", value.id),
-      (":touched-files", SExpList(value.touched.map(f => strToSExp(f.getAbsolutePath)))))
-  }
-
-  def toWF(value: SymbolDesignations): SExp = {
-    SExp.propList(
-      (":file", value.file),
-      (":syms",
-        SExpList(value.syms.map { s =>
-          SExpList(List(s.symType, s.start, s.end))
-        })))
-  }
-
-  def toWF(vmStatus: DebugVmStatus): SExp = {
-    vmStatus match {
-      case DebugVmSuccess() => SExp(
-        key(":status"), "success")
-      case DebugVmError(code, details) => SExp(
-        key(":status"), "error",
-        key(":error-code"), code,
-        key(":details"), details)
-    }
-  }
-
-  def toWF(method: MethodBytecode): SExp = {
-    SExp.propList(
-      (":class-name", method.className),
-      (":name", method.methodName),
-      (":signature", method.methodSignature.map(strToSExp).getOrElse('nil)),
-      (":bytecode", SExpList(method.byteCode.map { op =>
-        SExp(op.op, op.description)
-      })))
-  }
-
-  private def changeToWF(ch: FileEdit): SExp = {
-    SExp.propList(
-      (":file", ch.file.getCanonicalPath),
-      (":text", ch.text),
-      (":from", ch.from),
-      (":to", ch.to))
   }
 
 }
