+++ src/main/scala/org/ensime/protocol/messages.scala
@@ -1,0 +1,8 @@
+
+
+case class RPCRequest(data: RPCCall, callId: Int)
+
+sealed trait RPCCall
+
+case class ExecUndoRequest(id: Int) extends RPCCall
+
