+++ src/main/scala/org/ensime/server/Server.scala
@@ -2,14 +2,15 @@ package org.ensime.server
 
 import java.io._
 import java.net.{ ServerSocket, Socket, InetAddress }
+import akka.actor.{ ActorRef, Actor, Props, ActorSystem }
 import org.ensime.protocol._
 import org.ensime.util.WireFormat
 import org.ensime.config.Environment
-import scala.actors._
-import scala.actors.Actor._
+import org.slf4j.LoggerFactory
 import scala.util.Properties._
 
 object Server {
+  val log = LoggerFactory.getLogger("Server")
 
   def main(args: Array[String]): Unit = {
     def setFallbackProp(name: String, fallback: String): Unit = {
@@ -22,15 +23,12 @@ object Server {
 
     val (cacheDir, host, requestedPort) = if (args.length > 0) {
       // legacy interface
-      System.err.println(
-        "WARNING: org.ensime.server.Server now takes properties instead of arguments"
-      )
+      log.warn("WARNING: org.ensime.server.Server now takes properties instead of arguments")
       args match {
         case Array(a, b, c) => (
           new File(new File(a).getParentFile, ".ensime_cache"), b, c.toInt)
         case Array(portfile) => (
-          new File(new File(portfile).getParentFile, ".ensime_cache"),
-          "127.0.0.1", 0)
+          new File(new File(portfile).getParentFile, ".ensime_cache"), "127.0.0.1", 0)
         case _ =>
           throw new IllegalArgumentException("org.ensime.server.Server invoked incorrectly")
       }
@@ -40,6 +38,7 @@ object Server {
       propOrElse("ensime.requestport", "0").toInt
     )
 
+    val actorSystem = ActorSystem.create()
     require(!cacheDir.exists || cacheDir.isDirectory, cacheDir + " is not a valid cache directory")
     cacheDir.mkdirs()
 
@@ -53,20 +52,18 @@ object Server {
 
     writePort(cacheDir, actualPort)
 
-    val protocol = SwankProtocol
-    val project: Project = new Project(protocol)
-    project.start()
+    val protocol = new SwankProtocol
+    val project = new Project(protocol, actorSystem)
 
     try {
       while (true) {
         try {
           val socket = listener.accept()
-          println("Got connection, creating handler...")
-          val handler = new SocketHandler(socket, protocol, project)
-          handler.start()
+          log.info("Got connection, creating handler...")
+          val handler = actorSystem.actorOf(Props(classOf[SocketHandler], socket, protocol, project))
         } catch {
           case e: IOException =>
-            System.err.println("[ERROR] ENSIME Server: " + e)
+            log.error("ENSIME Server: ", e)
         }
       }
     } finally {
@@ -95,30 +92,34 @@ object Server {
   }
 }
 
-class SocketHandler(socket: Socket, protocol: Protocol, project: Project) extends Actor {
-  protocol.setOutputActor(this)
+case object SocketClosed
 
-  class SocketReader(socket: Socket, handler: SocketHandler) extends Actor {
-    val in = new BufferedInputStream(socket.getInputStream)
+class SocketReader(socket: Socket, protocol: Protocol, handler: ActorRef) extends Thread {
+  val in = new BufferedInputStream(socket.getInputStream)
 
-    def act() {
-      var running = true
-      try {
-        while (running) {
-          val msg: WireFormat = protocol.readMessage(in)
-          handler ! IncomingMessageEvent(msg)
-        }
-      } catch {
-        case e: IOException =>
-          System.err.println("Error in socket reader: " + e)
-          if (System.getProperty("ensime.explode.on.disconnect") != null) {
-            println("Tick-tock, tick-tock, tick-tock... boom!")
-            System.exit(-1)
-          } else exit('error)
+  override def run() {
+    try {
+      while (true) {
+        val msg: WireFormat = protocol.readMessage(in)
+        handler ! IncomingMessageEvent(msg)
       }
+    } catch {
+      case e: IOException =>
+        System.err.println("Error in socket reader: " + e)
+        if (System.getProperty("ensime.explode.on.disconnect") != null) {
+          println("Tick-tock, tick-tock, tick-tock... boom!")
+          System.exit(-1)
+        } else {
+          handler ! SocketClosed
+        }
     }
   }
+}
 
+class SocketHandler(socket: Socket, protocol: Protocol, project: Project) extends Actor {
+  protocol.setOutputActor(self)
+
+  val reader = new SocketReader(socket, protocol, self)
   val out = new BufferedOutputStream(socket.getOutputStream)
 
   def write(value: WireFormat) {
@@ -127,22 +128,21 @@ class SocketHandler(socket: Socket, protocol: Protocol, project: Project) extend
     } catch {
       case e: IOException =>
         System.err.println("Write to client failed: " + e)
-        exit('error)
+        context.stop(self)
     }
   }
 
-  def act() {
-    val reader = new SocketReader(socket, this)
-    this.link(reader)
+  override def preStart() {
     reader.start()
-    loop {
-      receive {
-        case IncomingMessageEvent(value: WireFormat) =>
-          project ! IncomingMessageEvent(value)
-        case OutgoingMessageEvent(value: WireFormat) =>
-          write(value)
-        case Exit(_: SocketReader, reason) => exit(reason)
-      }
-    }
+  }
+
+  override def receive = {
+    case IncomingMessageEvent(value: WireFormat) =>
+      project ! IncomingMessageEvent(value)
+    case OutgoingMessageEvent(value: WireFormat) =>
+      write(value)
+    case SocketClosed =>
+      System.err.println("Socket closed, stopping self")
+      context.stop(self)
   }
 }
